requirements.txt
django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0




